Update sql scripts for creating tables

+++ domain/src/main/resources/db/migration/basket/V20220421161946__Create_table_basket.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS basket
 (
     id              SERIAL PRIMARY KEY,
-    total_cost      INTEGER NOT NULL,
+    total_cost      REAL NOT NULL,
     person_id       INTEGER NOT NULL UNIQUE,
     CONSTRAINT fk_person
         FOREIGN KEY (person_id)
